+++ src/sc_crawler/alembic/versions/aeae56af8ca6_v0_3_4_convert_gpu_count_from_int_to_float.py
@@ -1,4 +1,4 @@
-"""v1.2.1 convert gpu_count from int to float
+"""v0.3.4 convert gpu_count from int to float
 
 Revision ID: aeae56af8ca6
 Revises: dad8a1f0f455