+++ src/sc_crawler/tables_scd.py
@@ -33,6 +33,9 @@ from .table_fields import (
 )
 from .tables import is_table
 
+# Foreign key fields from non-SCD tables are overridden as regular fields in SCD tables
+# because foreign key constraints cannot work properly when primary keys include `observed_at`.
+
 
 class Scd(ScModel):
     """Override the `observed_at` column to be primary key in SCD tables."""
@@ -54,7 +57,14 @@ class CountryScd(Scd, CountryBase, table=True):
 class VendorComplianceLinkScd(Scd, VendorComplianceLinkBase, table=True):
     """SCD version of .tables.VendorComplianceLink."""
 
-    pass
+    compliance_framework_id: str = Field(
+        primary_key=True,
+        description="Reference to the Compliance Framework.",
+    )
+    vendor_id: str = Field(
+        primary_key=True,
+        description="Reference to the Vendor.",
+    )
 
 
 class ComplianceFrameworkScd(Scd, ComplianceFrameworkBase, table=True):
@@ -66,55 +76,84 @@ class ComplianceFrameworkScd(Scd, ComplianceFrameworkBase, table=True):
 class VendorScd(Scd, VendorBase, table=True):
     """SCD version of .tables.Vendor."""
 
-    pass
+    country_id: str = Field(
+        description="Reference to the Country, where the Vendor's main headquarter is located.",
+    )
 
 
 class RegionScd(Scd, RegionBase, table=True):
     """SCD version of .tables.Region."""
 
-    pass
+    country_id: str = Field(
+        description="Reference to the Country, where the Region is located.",
+    )
+    vendor_id: str = Field(
+        primary_key=True,
+        description="Reference to the Vendor.",
+    )
 
 
 class ZoneScd(Scd, ZoneBase, table=True):
     """SCD version of .tables.Zone."""
 
-    pass
+    vendor_id: str = Field(
+        primary_key=True,
+        description="Reference to the Vendor.",
+    )
 
 
 class StorageScd(Scd, StorageBase, table=True):
     """SCD version of .tables.Storage."""
 
-    pass
+    vendor_id: str = Field(
+        primary_key=True,
+        description="Reference to the Vendor.",
+    )
 
 
 class ServerScd(Scd, ServerBase, table=True):
     """SCD version of .tables.Server."""
 
-    pass
+    vendor_id: str = Field(
+        primary_key=True,
+        description="Reference to the Vendor.",
+    )
 
 
 class ServerPriceScd(Scd, ServerPriceBase, table=True):
     """SCD version of .tables.ServerPrice."""
 
-    pass
+    vendor_id: str = Field(
+        primary_key=True,
+        description="Reference to the Vendor.",
+    )
 
 
 class StoragePriceScd(Scd, StoragePriceBase, table=True):
     """SCD version of .tables.StoragePrice."""
 
-    pass
+    vendor_id: str = Field(
+        primary_key=True,
+        description="Reference to the Vendor.",
+    )
 
 
 class TrafficPriceScd(Scd, TrafficPriceBase, table=True):
     """SCD version of .tables.TrafficPrice."""
 
-    pass
+    vendor_id: str = Field(
+        primary_key=True,
+        description="Reference to the Vendor.",
+    )
 
 
 class Ipv4PriceScd(Scd, Ipv4PriceBase, table=True):
     """SCD version of .tables.Ipv4Price."""
 
-    pass
+    vendor_id: str = Field(
+        primary_key=True,
+        description="Reference to the Vendor.",
+    )
 
 
 class BenchmarkScd(Scd, BenchmarkBase, table=True):
@@ -126,7 +165,14 @@ class BenchmarkScd(Scd, BenchmarkBase, table=True):
 class BenchmarkScoreScd(Scd, BenchmarkScoreBase, table=True):
     """SCD version of .tables.BenchmarkScore."""
 
-    pass
+    benchmark_id: str = Field(
+        primary_key=True,
+        description="Reference to the Benchmark.",
+    )
+    vendor_id: str = Field(
+        primary_key=True,
+        description="Reference to the Vendor.",
+    )
 
 
 tables_scd: List[SQLModel] = [o for o in globals().values() if is_table(o)]